# docs/api/console-api.openapi.yaml
openapi: 3.1.0
info:
  title: Batch Platform Console API
  version: 0.1.0
  description: >
    Frontend-facing contract for the batch console.
    This file is the machine-readable source of truth for UI integration,
    code generation, and mock server stubs.
    Whenever the API surface changes, update this file and console-api-protocol.md together.
servers:
  - url: http://localhost:8080
    description: Local console-api
...
security:
  - bearerAuth: []
  - consoleToken: []
paths:
...
  /api/console/query/audits:
    get:
      tags: [Queries]
      summary: Query audit logs
      operationId: queryAuditLogs
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Audit list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseConsoleAuditLogList'
  /api/console/query/execution-logs:
    get:
      tags: [Queries]
      summary: Query execution logs
      operationId: queryExecutionLogs
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Execution log list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseConsoleAuditLogList'
  /api/console/query/alerts:
    get:
      tags: [Queries]
      summary: Query alerts
      operationId: queryAlerts
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Alert list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseAlertEventList'
  /api/console/query/approvals:
    get:
      tags: [Queries]
      summary: Query approvals
      operationId: queryApprovals
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Approval list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseConsoleApprovalCommandList'
  /api/console/query/files:
    get:
      tags: [Queries]
      summary: Query file chains
      operationId: queryFiles
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: File chain list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseFileRecordList'
  /api/console/query/job-definitions:
    get:
      tags: [Queries]
      summary: Query job definitions
      operationId: queryJobDefinitions
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Job definition list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseJobDefinitionList'
  /api/console/query/outbox-retries:
    get:
      tags: [Queries]
      summary: Query outbox retry logs
      operationId: queryOutboxRetries
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Outbox retry list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseConsoleOutboxRetryLogList'
...
  /api/console/query/file-pipelines:
    get:
      tags: [Queries]
      summary: Query file pipelines
      operationId: queryFilePipelines
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: File pipeline list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseConsoleFilePipelineList'
  /api/console/query/file-pipeline-steps:
    get:
      tags: [Queries]
      summary: Query file pipeline steps
      operationId: queryFilePipelineSteps
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: File pipeline step list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseConsoleFilePipelineStepList'
  /api/console/query/file-dispatches:
    get:
      tags: [Queries]
      summary: Query file dispatch records
      operationId: queryFileDispatches
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: File dispatch list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseConsoleFileDispatchRecordList'
  /api/console/query/file-channels:
    get:
      tags: [Queries]
      summary: Query file channels
      operationId: queryFileChannels
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: File channel list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseConsoleFileChannelList'
  /api/console/query/file-arrival-groups:
    get:
      tags: [Queries]
      summary: Query file arrival groups
      operationId: queryFileArrivalGroups
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: File arrival group list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseFileArrivalGroupList'
  /api/console/query/file-errors:
    get:
      tags: [Queries]
      summary: Query file errors
      operationId: queryFileErrors
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: File error list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseFileErrorRecordList'
  /api/console/query/file-templates:
    get:
      tags: [Queries]
      summary: Query file templates
      operationId: queryFileTemplates
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: File template list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseConsoleFileTemplateList'
  /api/console/query/instances:
    get:
      tags: [Queries]
      summary: Query job instances
      operationId: queryJobInstances
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Job instance list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseJobInstanceList'
  /api/console/query/instances/{id}:
    get:
      tags: [Queries]
      summary: Query job instance detail
      operationId: queryJobInstanceDetail
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Job instance detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseConsoleJobInstanceResponse'
  /api/console/query/job-step-instances:
    get:
      tags: [Queries]
      summary: Query job step instances
      operationId: queryJobStepInstances
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Job step list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseJobStepInstanceList'
  /api/console/query/job-step-instances/{id}:
    get:
      tags: [Queries]
      summary: Query job step instance detail
      operationId: queryJobStepInstanceDetail
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Job step instance detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseConsoleJobStepInstanceResponse'
  /api/console/query/workflow-definitions:
    get:
      tags: [Queries]
      summary: Query workflow definitions
      operationId: queryWorkflowDefinitions
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Workflow definition list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseWorkflowDefinitionList'
  /api/console/query/workflow-nodes:
    get:
      tags: [Queries]
      summary: Query workflow nodes
      operationId: queryWorkflowNodes
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Workflow node list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseWorkflowNodeList'
  /api/console/query/workflow-edges:
    get:
      tags: [Queries]
      summary: Query workflow edges
      operationId: queryWorkflowEdges
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Workflow edge list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseWorkflowEdgeList'
  /api/console/query/workflow-runs:
    get:
      tags: [Queries]
      summary: Query workflow runs
      operationId: queryWorkflowRuns
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Workflow run list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseWorkflowRunList'
  /api/console/query/workflow-runs/{id}:
    get:
      tags: [Queries]
      summary: Query workflow run detail
      operationId: queryWorkflowRunDetail
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Workflow run detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseConsoleWorkflowRunResponse'
  /api/console/query/workflow-node-runs:
    get:
      tags: [Queries]
      summary: Query workflow node runs
      operationId: queryWorkflowNodeRuns
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageNoQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Workflow node run list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseWorkflowNodeRunList'
  /api/console/query/workflow-node-runs/{id}:
    get:
      tags: [Queries]
      summary: Query workflow node run detail
      operationId: queryWorkflowNodeRunDetail
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Workflow node run detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseConsoleWorkflowNodeRunResponse'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    consoleToken:
      type: apiKey
      in: header
      name: X-Console-Token
  parameters:
...
    TenantIdQuery:
      name: tenantId
      in: query
      required: false
      schema:
        type: string
...
    PageNoQuery:
      name: pageNo
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 1
    PageSizeQuery:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 20
...
  schemas:
...
    CommonResponseConsoleApprovalCommandList:
      allOf:
        - $ref: '#/components/schemas/CommonResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PageResponse'
...
    CommonResponseAlertEventList:
      allOf:
        - $ref: '#/components/schemas/CommonResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PageResponse'
...
    CommonResponseConsoleAuditLogList:
      allOf:
        - $ref: '#/components/schemas/CommonResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PageResponse'
...